+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
   * Shows different icon for category and scorecard in check result
+  * Fixes bug where scorecard section showed when an empty array returned
 
 ## [1.2.0] - 2023-11-03