+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+  * Nothing yet
+
+## [1.2.1] - 2023-11-06
+
 ### Changed
 
   * Shows different icon for category and scorecard in check result